+++ HISTORY.rst
@@ -9,6 +9,7 @@ Contributors to this version: Gabriel Rondeau-Genesse (:user:`RondeauG`), Juliet
 New features and enhancements
 ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
 * New functions ``diagnostics.properties_and_measures``, ``diagnostics.measures_heatmap`` and ``diagnostics.measures_improvement``. (:issue:`5`, :pull:`53`)
+* Add argument `resample_methods` to `xs.extract.resample`. (:issue:`57`, :pull:`57`)
 
 Breaking changes
 ^^^^^^^^^^^^^^^^
@@ -17,6 +18,7 @@ Breaking changes
 Internal changes
 ^^^^^^^^^^^^^^^^
 * Small bugfixes in aggregate.py (:pull:`55`, :pull:`56`).
+* Default method of `xs.extract.resample` now depends on frequency. (:issue:`57`, :pull:`57`).
 
 v0.3.0 (2022-08-23)
 -------------------